% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amidaKuji.R
\name{amidaKuji}
\alias{amidaKuji}
\title{Perform Amida Kuji for one player.}
\usage{
amidaKuji(playerNo, returnHistory = FALSE)
}
\arguments{
\item{playerNo}{No. of the player of interest. This argument should be integer.}

\item{returnHistory}{Return the history of Amida Kuji or not.}
}
\value{
\describe{
\item{playerNoFinal}{Where the player is at the final stage. Integer.}
\item{goalID}{The ID of the goal which corresponds to `playerNoFinal`}
\item{playerHistory}{The history of the path of the player. This shows the path that the player has passed.}
}
}
\description{
This function performs the Amida Kuji for one player of interest.
}
\examples{
### Create the base of Amida Kuji.
amidaKujiCreate(nVerticalLines = 12, nHorizontalLines = 30, noCross = TRUE)

### Show all the results of Amida Kuji
for(playerNo in 1:nVerticalLines){
cat(paste0("player: ", playerNames[playerNo], "\\n"))
print(amidaKuji(playerNo))
cat("--------------------------\\n")
}

}
