% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amidaKuji.R
\name{amidaKujiPlot}
\alias{amidaKujiPlot}
\title{Plot the Amida Kuji result}
\usage{
amidaKujiPlot(playerNo = NULL, col = 2)
}
\arguments{
\item{playerNo}{No. of the player of interest. This argument should be integer.
If NULL, Only the base of Amida Kuji will be shown.}

\item{col}{Color of the history of the path of the player of interest.}
}
\value{
Plot of Amida Kuji with some information.
}
\description{
This function shows the result of the Amida Kuji for one player of interest.
}
\examples{
### Create the base of Amida Kuji.
amidaKujiCreate(nVerticalLines = 12, nHorizontalLines = 30, noCross = TRUE)

### Plot the results of Amida Kuji
amidaKujiPlot(playerNo = 4, col = 2)

}
